chore: move record arena implementation to separate file
